requirements.txt
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
langchain==1.2.10
langchain-core==1.2.14
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.5
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
llama3-package==0.3.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
msgpack==1.1.2
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.3
notebook_shim==0.2.4
numba==0.64.0
numpy==2.4.2
ollama==0.6.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
platformdirs==4.9.2
pooch==1.9.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pywinpty==3.0.3; platform_system == "Windows"
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
Send2Trash==2.1.0
setuptools==82.0.0
six==1.17.0
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
stack-data==0.6.3
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xxhash==3.6.0
zstandard==0.25.0
fastapi
uvicorn
phonenumbers
crewai
elevenlabs
azure-cosmos
azure-communication-email
azure-communication-sms
azure-cognitiveservices-speech
azure-storage-blob
azure-search-documents
langchain-openai
# OpenAI Python SDK (api.openai.com + AzureOpenAI client in ai_service)
openai>=1.40.0
chromadb
sendgrid
pydantic-settings
structlog
